backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0,<5
psycopg2-binary==2.9.9
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.1
reportlab==4.0.9
aiofiles==23.2.1
